Mock_Test/requirements.txt
asgiref==3.8.1
Django==5.2
django-extensions==4.1
pillow==11.2.1
python-dotenv==1.1.0
sqlparse==0.5.3
markdown==3.8
